Add files via upload
I decided to do the product review for a PS5 for now. It can be changed though if you want to do something else.  However, you would need to alter the HTML and CSS code and find new icons and images. The home page is relatively finished. I'd recommend taking a look.